requirements.txt
openai
fitz
PyMuPDF
requests
beautifulsoup4
youtube-search-python
streamlit
selenium
python-dotenv
